Machine Learning with Cloud Computing/Practical 2/Part B/requirements.txt
matplotlib==3.8.3
pandas==2.2.0
python-dotenv==1.0.1
PyQt5==5.15.10 # Required for Microsoft Windows systems